More refactoring of FeatureDetectorFPN to be base class for both ObjectDetectorFPN and OpticalFlowFPN.
Usage of context structure and their pointers to allow for swapping of
src and dst when comparing frame to frame, as opposed to from a fixed
image into frames.

An interesting bug in OCV swaps the idx of Matches between trainIdx and
queryIndex, depending on which match routine is used and if the matcher
has been trained or not. Unpacking matches into key point homography
arrays needs to take this into account.. ugly!

Util routines to draw arrows and perform various geometry and
conversion utils